Drawing character on the level

+++ Game/core/src/org/academiadecodigo/hashtronauts/levels/gameObjects/ExitPoint.java
@@ -4,7 +4,7 @@ import com.badlogic.gdx.graphics.g2d.Sprite;
 
 public class ExitPoint extends GameObject {
 
-    public ExitPoint(Sprite sprite, float posX, float posY, int width, int height){
+    public ExitPoint(Sprite sprite, int posX, int posY, int width, int height){
         super(sprite, posX, posY, width, height);
     }
