+++ Game/core/src/org/academiadecodigo/hashtronauts/levels/gameObjects/GameObjectFactory.java
@@ -4,11 +4,11 @@ import com.badlogic.gdx.graphics.g2d.Sprite;
 
 public class GameObjectFactory {
 
-    public static Platform createPlatform(Sprite sprite, float posX, float posY, int width, int height){
+    public static Platform createPlatform(Sprite sprite, int posX, int posY, int width, int height){
         return new Platform(sprite, posX, posY, width, height);
     }
 
-    public static ExitPoint createExitPoint(Sprite sprite, float posX, float posY, int width, int height){
+    public static ExitPoint createExitPoint(Sprite sprite, int posX, int posY, int width, int height){
         return new ExitPoint(sprite, posX, posY, width, height);
     }
 }
